consolidate modal actions
