+++ app/actions/index.js
@@ -1,4 +1,3 @@
-import * as modals from './modalActions'
 import * as classroom from './classroomActions'
 import * as subject from './subjectActions'
 import * as exam from './examActions'
@@ -12,30 +11,30 @@ export const actionCreators = {
 	changeClassroomTab: classroom.changeClassroomTab,
 	handleClassData: classroom.handleClassData,
 	displayClassData: classroom.displayClassData,
-	roomModalDisplay: modals.roomModalDisplay,
-	deleteRoom: modals.deleteRoom,
-	updateRoom: modals.updateRoom,
+	roomModalDisplay: classroom.roomModalDisplay,
+	deleteRoom: classroom.deleteRoom,
+	updateRoom: classroom.updateRoom,
 	openClassList: subject.openClassList,
 	showSubject: subject.showSubject,
 	addNewSubject: subject.addNewSubject,
 	getSubjectData: subject.getSubjectData,
-	updateSubject: modals.updateSubject,
-	subjectModalDisplay: modals.subjectModalDisplay,
-	deleteSingleSubject: modals.deleteSingleSubject,
+	updateSubject: subject.updateSubject,
+	subjectModalDisplay: subject.subjectModalDisplay,
+	deleteSingleSubject: subject.deleteSingleSubject,
 	addNewExam: exam.addNewExam,
 	getSelectedSubject: exam.getSelectedSubject,
 	openClassDropdownList: exam.openClassDropdownList,
 	displayExamData: exam.displayExamData,
 	showSingleExam: exam.showSingleExam,
-	deleteSingleExam: modals.deleteSingleExam,
-	updateExam: modals.updateExam,
+	deleteSingleExam: exam.deleteSingleExam,
+	updateExam: exam.updateExam,
 	addNewStudent: student.addNewStudent,
 	getStudents: student.getStudents,
 	showStudentModal: student.showStudentModal,
 	openStudenGraph: student.openStudenGraph,
 	openStudenSubjectGraph: student.openStudenSubjectGraph,
-	deleteSingleStudent: modals.deleteSingleStudent,
-	updateStudent: modals.updateStudent,
+	deleteSingleStudent: student.deleteSingleStudent,
+	updateStudent: student.updateStudent,
 	displayGradeData: grade.displayGradeData,
 	openGradeClassList: grade.openGradeClassList,
 	updateGrade: grade.updateGrade,
