+++ app/actions/studentActions.js
@@ -5,7 +5,12 @@ import {
 	DISPLAY_STUDENT_GRAPH,
 	DISPLAY_STUDENT_SUBJECT_GRAPH
 } from '../constants/actionTypes'
-import { addNewStudentData, getAllStudents } from '../database/studentCollection'
+import {
+	addNewStudentData,
+	getAllStudents,
+	deleteStudent,
+	updateStudentData
+} from '../database/studentCollection'
 
 export const addNewStudent = event => async dispatch => {
 	event.preventDefault()
@@ -79,4 +84,46 @@ export const openStudenSubjectGraph = event => dispatch => {
 		type: DISPLAY_STUDENT_SUBJECT_GRAPH,
 		payload: subject
 	})
+}
+
+export const updateStudent = event => async dispatch => {
+	event.preventDefault()
+	const genderIndex = event.target.gender.selectedIndex
+	const studentData = {
+		firstname: event.target.firstname.value,
+		lastname: event.target.lastname.value,
+		gender: event.target.gender.options[genderIndex].getAttribute('data-id'),
+		classroom: event.target.classroom.value,
+		id: event.target.studentId.getAttribute('data-id')
+	}
+
+	const students = await updateStudentData(studentData)
+
+	dispatch({
+		type: GET_SINGLE_STUDENT,
+		payload: studentData.getAttribute('data-id')
+	})
+
+	if (students.length > 0) {
+		dispatch({
+			type: GET_ALL_STUDENTS,
+			payload: { students }
+		})
+	}
+}
+
+export const deleteSingleStudent = event => async dispatch => {
+	const studentId = event.target.getAttribute('data-id')
+	const students = await deleteStudent(studentId)
+
+	dispatch({
+		type: GET_SINGLE_STUDENT,
+		payload: studentId
+	})
+	if (students.length > 0) {
+		dispatch({
+			type: GET_ALL_STUDENTS,
+			payload: { students }
+		})
+	}
 }
